renamed spi/i2c to generic linux since they're not platform specific and updated busio to include it when possible
