+++ src/adafruit_blinka/microcontroller/generic_linux/spi.py
